completed ai insights route
