+++ apartment_portal_api/apartment_portal_api/Data/UnitOfWork.cs
@@ -2,15 +2,12 @@
 using apartment_portal_api.Data.Repositories;
 using apartment_portal_api.Models;
 using apartment_portal_api.Models.Guests;
-using apartment_portal_api.Models.Insights;
 using apartment_portal_api.Models.ParkingPermits;
-using apartment_portal_api.Models.Issues;
 using apartment_portal_api.Models.IssueTypes;
-using apartment_portal_api.Models.Packages;
 using apartment_portal_api.Models.Statuses;
 using apartment_portal_api.Models.Users;
 using apartment_portal_api.Models.UnitUsers;
-using apartment_portal_api.Services;
+using apartment_portal_api.Services.AIService;
 
 namespace apartment_portal_api.Data;
