+++ apartment_portal_api/apartment_portal_api/Services/AIService/Candidate.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace apartment_portal_api.Services.AIService;
+
+public class Candidate
+{
+    [JsonPropertyName("content")] public Content Content { get; set; } = new();
+}
