+++ apartment_portal_api/apartment_portal_api/Services/AIService/Content.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace apartment_portal_api.Services.AIService;
+
+public class Content
+{
+    [JsonPropertyName("parts")] public List<Parts> Parts { get; set; } = new();
+}
