+++ apartment_portal_api/apartment_portal_api/Services/AIService/GeminiAPIResponse.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace apartment_portal_api.Services.AIService;
+
+public class GeminiApiResponse
+{
+    [JsonPropertyName("candidates")] public List<Candidate> Candidates { get; set; } = new();
+}
