+++ apartment_portal_api/apartment_portal_api/Services/AIService/Parts.cs
@@ -1,0 +1,13 @@
+﻿using System.Text.Json.Serialization;
+
+namespace apartment_portal_api.Services.AIService;
+
+public class Parts
+{
+    [JsonPropertyName("text")] public string Text { get; set; }
+
+    public Parts(string text)
+    {
+        Text = text;
+    }
+}
